requirements.txt
asgiref==3.5.0
autopep8==1.7.0
backports.zoneinfo;python_version<"3.9.15"
build==0.8.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.1.3
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
django-extensions==3.1.5
django-heroku==0.3.1
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pep517==0.13.0
Pillow==9.2.0
pip-tools==6.8.0
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycryptodome==3.15.0
PyJWT==2.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
rave-python==1.2.16
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sentry-sdk==1.5.4
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
twilio==7.14.1
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.14.1
whitenoise==5.3.0
yapf==0.32.0
